commit messages are safe now
