+++ source/Calamari/ArgoCD/Conventions/ArgoCommitToGitConfigFactory.cs
@@ -24,10 +24,9 @@ namespace Calamari.ArgoCD.Conventions
             var recursive = deployment.Variables.GetFlag(SpecialVariables.Git.Recursive, false);
             
             var requiresPullRequest = RequiresPullRequest(deployment);
-            
-            //These two are problematic as they MAY be sensitive
-            var summary = GetNonSensitiveVariable(SpecialVariables.Git.CommitMessageSummary);
-            var description = GetNonSensitiveVariable(SpecialVariables.Git.CommitMessageDescription) ?? string.Empty;
+
+            var summary = deployment.GetMandatoryNonSensitiveVariable(SpecialVariables.Git.CommitMessageSummary);
+            var description = deployment.GetNonSensitiveVariable(SpecialVariables.Git.CommitMessageDescription) ?? string.Empty;
             
             var repositoryIndexes = deployment.Variables.GetIndexes(SpecialVariables.Git.Index);
             log.Info($"Found the following repository indicies '{repositoryIndexes.Join(",")}'");
@@ -50,35 +49,6 @@ namespace Calamari.ArgoCD.Conventions
         bool RequiresPullRequest(RunningDeployment deployment)
         {
             return OctopusFeatureToggles.ArgoCDCreatePullRequestFeatureToggle.IsEnabled(deployment.Variables) && deployment.Variables.Get(SpecialVariables.Git.CommitMethod) == SpecialVariables.Git.GitCommitMethods.PullRequest;
-        }
-
-        string GetNonSensitiveVariable(string variableName, RunningDeployment deployment)
-        {
-            var result = deployment.Variables.Get(variableName, out var error);
-            if (!string.IsNullOrEmpty(error))
-            {
-                var message = $"Unable to evaluate '{variableName}'. It may reference missing or sensitive values.";
-                throw new InvalidOperationException(message);
-            }
-
-            return result;
-        }
-        
-        string GetMandatoryNonSensitiveVariable(string variableName, RunningDeployment deployment)
-        {
-            var result = deployment.Variables.Get(variableName, out var error) ?? null;
-            if (string.IsNullOrWhiteSpace(result))
-            {
-                throw new CommandException($"Variable {variableName} was not supplied");
-            }
-            
-            if (!string.IsNullOrEmpty(error))
-            {
-                var message = $"Unable to evaulate '{variableName}'. It may reference missing or sensitive values.";
-                throw new InvalidOperationException(message);
-            }
-
-            return result;
         }
     }
 }
